branch plugin: do not set mac addr for branch vlan
1. Got rid of the code that was setting the branch mac address for the vlan
device. Doing so would result in incorrect n/w configuration where the
VM would not be able to retrieve its IPv4 address. This change fixes
that bug.
2. Switch the name of the bridge to 'br' prefix from 'tapbr' prefix.
Added code comments for reasons to do so. This is only a WIP change and
needs to be handled differently in the future (as per code comments).

cr https://code.amazon.com/reviews/CR-2506647