fix pat-veth-tap setup
When pat netns was being connected to target netns (where the tap and
tap bridge interfaces are being created) with a veth pair, the tap
interface would fail to get the ip address assignment.

This was mostly because the interfaces being created in these namespaces
were not being brought up (were in DOWN state) and because the mac
address for the vlan device was not being overwritten (due to an
erroneous change in 848e4ce). This commit fixes both of these issues.

cr https://code.amazon.com/reviews/CR-2609290