pat config: Added cleanupPATNetNS flag
Today, the PAT network namespace gets deleted when the last veth link
created in the namesapce gets deleted during the 'Del()' call. The
cleanupPATNetNS flag can be used to override that logic. With
this commit, the behavior is changed so that the pat network namespaces
id deleted only when this flag is set to true on top of the already
existing last veth pair deleted check.

cr https://code.amazon.com/reviews/CR-2655487