bugfix: longer than allowed veth pair name length
If the container id passed to the pat plugin is too long, veth pair
creation used to fail because of restriction of 15 bytes on the names of
IF devices in linux.

This change modifies the behavior of the plugin to create random veth
pair names in such a scenario. A corresponding change in the Del()
operation is to look up veth pair device by type instead of by name.

cr https://code.amazon.com/reviews/CR-3744638