vpc-branch-eni plugin: use gateway ip address if specified in netconfig
Previously the branchGatewayIPAddress field in NetConfig of vpc-branch-eni can be provided but was never used, and the gateway ip was always inferred from the subnet. Existing behavior of ecs task networking is to supply the gateway ip address to ecs-eni plugin with the value specified in acs payload, and that will be the gateway ip to use. To use vpc-branch-eni plugin in eni trunking, we need to maintain the same behavior. This commit makes the necessary changes for that in the plugin.