vpc-shared-eni: Remove call to get Pod from API Server in DEL command
We don't need to query the API Server on DEL command as we get the
Endpoint details using the container ID. We only query the API Server
on ADD comand to get the IP Address from the Pod's Annotation.

Note: IP Address information will be required for EKS Linux node which
are not supported yet.

Resolves: #42