Add vpc-tunnel plugin
The `vpc-tunnel` CNI plugin sets up and manages GENEVE interfaces for
communication over a network tunnel. The plugin can be invoked in
2 modes - `geneve` and `tap`. In the `geneve` mode, the plugin sets up
and configures the GENEVE interface on the host. In the `tap` mode, it
sets up a bridge and a tap interface connecting the GENEVE interface
with processes like a MicroVM.