add HNS V2 support for vpc-shared-eni Windows CNI plugin
The containers launched using containerd are launched into new HCN network namespaces on Windows. The networking for HCN Namespaces is performed using the HNS V2 apis. The changes introduced in this commit allows the plugin to setup networking even for such cases. The backward compatibility is maintained so that HNS V1 use cases remain supported.