windows: enhance vpc-bridge plugin to use HNS V2 APIs by default
Presently, in the `vpc-bridge` plugin, we use HNS V1 APIs to create the HNS network and the HNS endpoint. Once the endpoint is created, we use either V1 or V2 APIs for attaching the endpoint to the container.

Given that V2 is the way forward where Microsoft supports newer features, we are converting our plugins to be V2 first. In this commit, we make the following changes-
- We use HNS V2 to create the network
- We use HNS V2 to create the HNS endpoint

After that, we will use either V1 or V2 to attach the endpoint to our container/namespace depending upon the runtime.